Full seeding

+++ api/DbEntities/FoodNutrient.cs
@@ -4,7 +4,7 @@ namespace api.DbEntities;
 
 public class FoodNutrient : BaseEntity
 {
-    public required int Amount { get; set; }
+    public required float Amount { get; set; }
     public required Food Food { get; set; }
     public required Nutrient Nutrient { get; set; }
 }
